+++ src/index.ts
@@ -5,34 +5,13 @@ import {
 } from './__generated__/GPUBenchmark';
 
 // Modules
+import { cleanEntryString } from './cleanEntryString';
+import { cleanRendererString } from './cleanRendererString';
 import { getBenchmarkByPercentage } from './getBenchmarkByPercentage';
 import { isDesktop, isMobile, isTablet } from './getBrowserType';
+import { getEntryVersionNumber } from './getEntryVersionNumber';
 import { getWebGLUnmaskedRenderer } from './getWebGLUnmaskedRenderer';
 import { isWebGLSupported } from './isWebGLSupported';
-
-const cleanEntryString = (entryString: string): string =>
-  entryString
-    .toLowerCase() // Lowercase all for easier matching
-    .split('- ')[1] // Remove prelude score (`3 - `)
-    .split(' /')[0]; // Reduce 'apple a9x / powervr series 7xt' to 'apple a9x'
-
-const getEntryVersionNumber = (entryString: string): string => entryString.replace(/[\D]/g, ''); // Grab and concat all digits in the string
-
-const cleanRendererString = (rendererString: string): string => {
-  let cleanedRendererString = rendererString.toLowerCase();
-
-  // Strip off ANGLE and Direct3D version
-  if (cleanedRendererString.includes('angle (') && cleanedRendererString.includes('direct3d')) {
-    cleanedRendererString = cleanedRendererString.replace('angle (', '').split(' direct3d')[0];
-  }
-
-  // Strip off the GB amount (1060 6gb was being concatenated to 10606 and because of it using the fallback)
-  if (cleanedRendererString.includes('nvidia') && cleanedRendererString.includes('gb')) {
-    cleanedRendererString = cleanedRendererString.split(/\dgb/)[0];
-  }
-
-  return cleanedRendererString;
-};
 
 export interface IGetGPUTier {
   mobileBenchmarkPercentages?: number[];
